requirements.txt
# AeroZeka - Filo Atama
# Python 3.8+

# Gerçek uçuş verisi (API key gerektirmez)
FlightRadarAPI>=1.3.34

# Harita bileşeni
tkintermapview>=1.20

# Uçak thumbnail ve placeholder resimler
Pillow>=9.0

# tkintermapview / FlightRadarAPI ortak bağımlılıkları (genelde otomatik kurulur)
# requests, customtkinter, geocoder
